Moved functions to 'socket_functions.py' to help clean up the main scripts
