+++ client.py
@@ -2,92 +2,10 @@
 # client.py
 
 import sys
-import socket
-import time
 import getopt
 from socket_helper import fresh_client_socket
+from socket_functions import send_latency, receive_latency, calc_throughput
 
-
-"""
-    Methods start here
-"""
-
-def send_latency(s_lat_len, s_lat_sock):
-    i = 0
-
-    lat_array = bytearray()
-    lat_array.append(1)
-    i += 1
-
-    while i < s_lat_len - 1:
-        lat_array.append(0)
-        i += 1
-
-    if s_lat_len != 1:
-        lat_array.append(255)
-
-    s_time = time.clock()
-    amount_sent = 0
-    while amount_sent < len(lat_array):
-        sent = s_lat_sock.send(lat_array[amount_sent:])
-        amount_sent += sent
-
-    s_lat_sock.shutdown(socket.SHUT_RD)
-    return s_time
-
-def receive_latency(r_lat_len, r_lat_sock):
-    received_bytes = []
-
-    data = r_lat_sock.recv(r_lat_len)
-    while data != '':
-        received_bytes.append(bytes(data))
-        data = r_lat_sock.recv(r_lat_len)
-
-    r_time = time.clock()
-    r_lat_sock.close()
-    return r_time
-
-def calc_throughput(tp_size, tp_sock):
-    i = 0
-
-    tp_array = bytearray()
-    tp_array.append(2)
-
-    while i < tp_size - 1:
-        tp_array.append(0)
-        i += 1
-
-    if tp_size != 1:
-        tp_array.append(255)
-        i += 1
-
-    s_time = time.clock()
-    amount_sent = 0
-    while amount_sent < len(tp_array):
-        sent = tp_sock.send(tp_array[amount_sent:])
-        amount_sent += sent
-
-    tp_sock.shutdown(socket.SHUT_RD)
-
-    reply_bytes = []
-
-    data = tp_sock.recv(4096)
-    while data != '':
-        reply_bytes.append(data)
-        data = tp_sock.recv(4096)
-
-    r_time = time.clock()
-    tp_sock.close()
-
-    tx_time = (r_time + s_time) / 2
-
-    tx_rate = (tp_size / 1000) / tx_time
-
-    return tx_rate
-
-"""
-    This is where the program script starts
-"""
 
 PORT = 2694  # The port used for the socket
 HOST_SERVER = None
