+++ server.py
@@ -5,60 +5,7 @@ import socket
 import getopt
 import sys
 from socket_helper import fresh_server_socket
-
-def test_latency(lat_data, sock):
-    data_len = len(lat_data)
-    lat_reply = bytearray()
-
-    i = 0
-    lat_reply.append(1)
-
-    if data_len != 1:
-        i += 1
-        while i < data_len - 1:
-            lat_reply.append(0)
-            i += 1
-
-        lat_reply.append(255)
-        i += 1
-
-    print "Sending latency reply..."
-    amount_sent = 0
-    while amount_sent < len(lat_reply):
-        sent = sock.send(lat_reply[amount_sent:])
-        amount_sent += sent
-        print "** Sent " + str(sent) + " of " + str(len(lat_reply)) + "..."
-
-    print "Reply sent..."
-
-def test_throughput(tp_data, tp_sock):
-    # Send the data back...
-    print "Sending throughput reply..."
-
-    data_len = len(tp_data)
-    tp_reply = bytearray()
-
-    i = 0
-
-    tp_reply.append(2)
-
-    if data_len != 1:
-        i += 1
-        while i < data_len - 1:
-            tp_reply.append(0)
-            i += 1
-
-        tp_reply.append(255)
-        i += 1
-
-    amount_sent = 0
-    while amount_sent < len(tp_reply):
-        sent = tp_sock.send(tp_reply[amount_sent:])
-        amount_sent += sent
-        print "** Sent " + str(sent) + " of " + str(len(tp_reply)) + "..."
-
-    tp_sock.close()
-    print "Throughput reply sent..."
+from socket_functions import test_latency, test_throughput
 
 
 port = 2694 # The port number assigned in class
